+++ CHANGELOG.md
@@ -5,13 +5,13 @@ CHANGELOG
 * No longer using separateCSS with jspm.bundle
 * jspm 0.17
 * gulp 4.0
-* font-awesome
+* Font Awesome
 * gulp build now builds directly in dist directory
 * Bootstrap 4.0-alpha.2
 * angular-ui-bootstrap@^1.3.2
+* Using TypeScript as default SystemJS transpiler
 
 --------------------
 
 ## 1.0.0-alpha (2016-04-28)
 _(none)_
-